refactor(ripgrep): split into directory with separate modules
Split ripgrep.ts into a directory structure for better organization:
- binary.ts: Platform config, download, extract, filepath
- schema.ts: Zod schemas for ripgrep JSON output
- tree.ts: TreeNode interface and tree helpers (build, sort, truncate, render)
- index.ts: Public API (files, tree, search) and re-exports
